+++ fail/fail.go
@@ -48,6 +48,7 @@ const (
 	ErrDivisionByZero          = "division by zero - divisor cannot be zero"
 	ErrSomeDirsOnlyInTemplates = "@use, @insert, @reserve, @component only allowed in templates"
 	ErrGlobalFuncMissing       = "global function '%s' not found"
+	ErrProperyOnNonObject      = "cannot get propery '%s' from a non-object type '%s'"
 
 	// Functions
 	ErrNoFuncForThisType  = "function '%s' not available for type '%s'"